+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/NetworkNodeExporter.java
@@ -3,6 +3,7 @@ package com.raoulvdberge.refinedstorage.apiimpl.network.node;
 import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.api.util.Action;
 import com.raoulvdberge.refinedstorage.api.util.IComparer;
+import com.raoulvdberge.refinedstorage.apiimpl.network.node.cover.CoverManager;
 import com.raoulvdberge.refinedstorage.inventory.ItemHandlerBase;
 import com.raoulvdberge.refinedstorage.inventory.ItemHandlerFluid;
 import com.raoulvdberge.refinedstorage.inventory.ItemHandlerListenerNetworkNode;
@@ -17,17 +18,19 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.minecraftforge.common.util.Constants;
 import net.minecraftforge.fluids.Fluid;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.ItemHandlerHelper;
 
-public class NetworkNodeExporter extends NetworkNode implements IComparable, IType {
+public class NetworkNodeExporter extends NetworkNode implements IComparable, IType, ICoverable {
     public static final String ID = "exporter";
 
     private static final String NBT_COMPARE = "Compare";
     private static final String NBT_TYPE = "Type";
+    private static final String NBT_COVERS = "Covers";
 
     private ItemHandlerBase itemFilters = new ItemHandlerBase(9, new ItemHandlerListenerNetworkNode(this));
     private ItemHandlerFluid fluidFilters = new ItemHandlerFluid(9, new ItemHandlerListenerNetworkNode(this));
@@ -36,6 +39,8 @@ public class NetworkNodeExporter extends NetworkNode implements IComparable, ITy
 
     private int compare = IComparer.COMPARE_NBT | IComparer.COMPARE_DAMAGE;
     private int type = IType.ITEMS;
+
+    private CoverManager coverManager = new CoverManager(this).setCanPlaceCoversOnFace(false);
 
     private int filterSlot;
 
@@ -149,12 +154,6 @@ public class NetworkNodeExporter extends NetworkNode implements IComparable, ITy
         markDirty();
     }
 
-    @Override
-    public void read(NBTTagCompound tag) {
-        super.read(tag);
-
-        StackUtils.readItems(upgrades, 1, tag);
-    }
 
     @Override
     public String getId() {
@@ -166,6 +165,8 @@ public class NetworkNodeExporter extends NetworkNode implements IComparable, ITy
         super.write(tag);
 
         StackUtils.writeItems(upgrades, 1, tag);
+
+        tag.setTag(NBT_COVERS, coverManager.writeToNbt());
 
         return tag;
     }
@@ -181,6 +182,17 @@ public class NetworkNodeExporter extends NetworkNode implements IComparable, ITy
         StackUtils.writeItems(fluidFilters, 2, tag);
 
         return tag;
+    }
+
+    @Override
+    public void read(NBTTagCompound tag) {
+        super.read(tag);
+
+        StackUtils.readItems(upgrades, 1, tag);
+
+        if (tag.hasKey(NBT_COVERS)) {
+            coverManager.readFromNbt(tag.getTagList(NBT_COVERS, Constants.NBT.TAG_COMPOUND));
+        }
     }
 
     @Override
@@ -223,5 +235,10 @@ public class NetworkNodeExporter extends NetworkNode implements IComparable, ITy
     @Override
     public IItemHandler getFilterInventory() {
         return getType() == IType.ITEMS ? itemFilters : fluidFilters;
+    }
+
+    @Override
+    public CoverManager getCoverManager() {
+        return coverManager;
     }
 }