+++ src/main/java/com/raoulvdberge/refinedstorage/block/BlockExternalStorage.java
@@ -15,7 +15,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 import javax.annotation.Nullable;
-import java.util.ArrayList;
 import java.util.List;
 
 public class BlockExternalStorage extends BlockCable {
@@ -32,7 +31,7 @@ public class BlockExternalStorage extends BlockCable {
 
     @Override
     public List<AxisAlignedBB> getCollisionBoxes(TileEntity tile, IBlockState state) {
-        List<AxisAlignedBB> boxes = new ArrayList<>();
+        List<AxisAlignedBB> boxes = super.getCollisionBoxes(tile, state);
 
         switch (state.getValue(getDirection().getProperty())) {
             case NORTH: