+++ src/main/java/com/raoulvdberge/refinedstorage/proxy/ProxyClient.java
@@ -372,7 +372,7 @@ public class ProxyClient extends ProxyCommon {
             if (model.getResourceDomain().equals(RS.ID)) {
                 if (model.getResourcePath().equals("pattern")) {
                     e.getModelRegistry().putObject(model, new BakedModelPattern(e.getModelRegistry().getObject(model)));
-                } else if (model.getResourcePath().equals("cable")) {
+                } else if (model.getResourcePath().equals("cable") || model.getResourcePath().equals("exporter") || model.getResourcePath().equals("importer") || model.getResourcePath().equals("external_storage")) {
                     e.getModelRegistry().putObject(model, new BakedModelCableCover(e.getModelRegistry().getObject(model)));
                 } else if (model.getResourcePath().equals("cover")) {
                     e.getModelRegistry().putObject(model, new BakedModelCover(e.getModelRegistry().getObject(model), null));